INTERLOK-2345 Added null protections
So, it appears that travis doesn't have the "same" environment
as the local machine (or even our jenkins instance)... and an
NPE is being throw.
The likely culprit is that  dir.listFiles() returns null; which
means that the directory doesn't exist... so we add some protection
and some explicit tests.
travis debugging isn't helpful.